+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - Env: added BITTYTAX_DATA_DIR variable which specifies the location for storing the .bittytax folder.
 - Accounting tool: tax rates and allowance for 2026/27.
 - Conversion tool: added newest_first to DataParser to control the parsing order.
+- KuCoin parser: handle empty reports with "No matching records found".
 ### Changed
 - Config: fiat_income to True.
 - Price tool: CoinDesk API deprecated.